move Mutations::Command monkeypatch
