+++ server/app/boot.rb
@@ -12,7 +12,7 @@ require 'roda'
 require 'mongoid'
 require 'json'
 require 'mutations'
-require_relative '../lib/mutations_command'
+require_relative '../lib/mutations/command_errors'
 require 'logger'
 require 'msgpack'
 require 'tilt/jbuilder.rb'
