+++ server/lib/mutations/command_errors.rb
